refactor: move the group_library to the inside of the hub repo
With this change we can in theory have multi-platform libraries in the
dependency cycle and use the pip hub repo for the dependencies. With
this we can finally make use of multi-platform whl_libraries and have
the correct transitive dependency closure when we reuse the same
`whl_library` for platforms targeting different python versions and
os/cpu tuples.